refactor: Rename etl.collections as etl.collection
