+++ apps/wizard/app_pages/explorer_edit.py
@@ -4,7 +4,7 @@ from io import StringIO
 
 import streamlit as st
 
-from etl.collections.explorer_legacy import ExplorerLegacy
+from etl.collection.explorer_legacy import ExplorerLegacy
 
 # PAGE CONFIG
 st.set_page_config(
