+++ docs/guides/data-work/mdims.md
@@ -16,7 +16,7 @@ These have a similar structure to regular steps and are run with the `--export` 
 etlr export://explorers/minerals/latest/minerals --export
 ```
 
-Instead of creating a `Dataset` object, these create `MDIM` or `Explorer` objects and save them to the database. Both `MDIM` and `Explorer` are based on the same underlying data structure (class `Collection`, see in [`etl.collections.model`](https://github.com/owid/etl/blob/master/etl/collections/model.py)), but have minor differences due to legacy reasons.
+Instead of creating a `Dataset` object, these create `MDIM` or `Explorer` objects and save them to the database. Both `MDIM` and `Explorer` are based on the same underlying data structure (class `Collection`, see in [`etl.collection.model`](https://github.com/owid/etl/blob/master/etl/collection/model.py)), but have minor differences due to legacy reasons.
 
 !!! question "Why do we need both?"
     We understand MDIMs as a more powerful version of explorers, and will likely replace them in the long-run from a technical perspective. MDIMs still are still missing some key features to fully be able to replace explorers.
@@ -38,7 +38,7 @@ For details on how to implement an explorer step, first read this section and th
     - [Explorers :material-arrow-right:](https://github.com/owid/etl/blob/master/etl/steps/export/explorers/)
 
 ### A simple MDIM
-MDIMs can be crafted in various ways, going from totally manual to fully automated. The most common way is to use a combination of manual and automated steps. The automated steps are usually generated by helper functions from the `etl.collections` module.
+MDIMs can be crafted in various ways, going from totally manual to fully automated. The most common way is to use a combination of manual and automated steps. The automated steps are usually generated by helper functions from the `etl.collection` module.
 
 Below is a simple MDIM configuration file (modified for demo purposes):
 
@@ -83,7 +83,7 @@ views:
 This example is fairly simple, and is entirely manual. The corresponding step script could simply look like:
 
 ```python
-from etl.collections import multidim
+from etl.collection import multidim
 from etl.helpers import PathFinder
 
 # Get paths and naming conventions for current step.
@@ -108,7 +108,7 @@ In this example, we note that we can group together indicators from any dataset.
     Proper documentation of all available options will be available soon.
 
 ### Automated MDIMs
-MDIMs are very suitable for those datasets that have a lot of dimensions in them. For these, one can leverage functions like `etl.collections.multidim.expand_config`, which programmatically generates all possible views from an indicator (or multiple ones) in a table.
+MDIMs are very suitable for those datasets that have a lot of dimensions in them. For these, one can leverage functions like `etl.collection.multidim.expand_config`, which programmatically generates all possible views from an indicator (or multiple ones) in a table.
 
 !!! question "What does it mean for a dataset to have dimensions?"
     Datasets with dimensions are those that have additional index (e.g. `age`) in Garden, beyond the standard ones (`year` or `date`, and `country`).
@@ -192,7 +192,7 @@ def run() -> None:
 
 ```
 
-You can also combine manually defined views with generated ones. See the `etl.collections.multidim` module for available helper functions or refer to examples from `etl/steps/export/multidim/`.
+You can also combine manually defined views with generated ones. See the `etl.collection.multidim` module for available helper functions or refer to examples from `etl/steps/export/multidim/`.
 
 
 ### Defining view configurations
@@ -259,7 +259,7 @@ etlr export://multidim/energy/latest/energy_prices --export
 and check out the preview at: http://staging-site-my-branch/admin/multi-dims
 
 ### Experimental tooling
-There is on going work in `etl.collections.beta`. In there you can find interesting methods like `combine_explorers`, which lets you combine two explorer objects into one. Since this works also with MDIMs, it can be extremely useful to combine MDIMs automatically-generated from different tables.
+There is on going work in `etl.collection.beta`. In there you can find interesting methods like `combine_explorers`, which lets you combine two explorer objects into one. Since this works also with MDIMs, it can be extremely useful to combine MDIMs automatically-generated from different tables.
 
 ## Nuances of creating an Explorer
 While Explorer and MDIM steps are almost identical, there are some subtle implementation differences. We hope that this is just a temporary situation and that we will be able to merge them into a single step in the future.
@@ -272,7 +272,7 @@ Some key differences are:
 - **View config has different fields:** MDIMs rely on our standard grapher configs. Instead, Explorers rely on a custom configuration. Fields available are equivalent to the columns that are available in the `grapher` table of a legacy explorer config in admin.
 - **Indicator display config**: Explorers also rely on a custom configuration. Fields available are equivalent to the columns that are available in the `columns` table of a legacy explorer config in admin.
 
-Also, instead of relying on `etl.collections.multidim`, you should instead use the functions from `etl.collections.explorer`.
+Also, instead of relying on `etl.collection.multidim`, you should instead use the functions from `etl.collection.explorer`.
 
 !!! note "There is no schema for explorers"
