+++ etl/steps/export/explorers/animal_welfare/latest/animal_welfare.py
@@ -2,7 +2,7 @@
 
 from copy import deepcopy
 
-from etl.collections.explorer import expand_config
+from etl.collection.explorer import expand_config
 from etl.helpers import PathFinder
 
 # Get paths and naming conventions for current step.
