+++ etl/steps/export/explorers/faostat/latest/global_food.py
@@ -6,7 +6,7 @@ from typing import Any, Callable, Dict, Optional
 from owid.catalog.utils import underscore
 from structlog import get_logger
 
-from etl.collections.explorer import combine_config_dimensions, expand_config
+from etl.collection.explorer import combine_config_dimensions, expand_config
 from etl.helpers import PathFinder
 
 # Initialize log.
