+++ etl/steps/export/explorers/un/latest/un_wpp.py
@@ -7,13 +7,13 @@ Strategy:
     - While some of the metadata is inherited from Garden/Grapher, some of it is set manually in the YAML files or programmatically once explorers are created.
     - Most of the explorers are created programmatically, with slight edits coming from YAML files (un_wpp.config.yml and un_wpp.sex_ratio.yml).
     - In addition, some views, were created using manual configuration (see un_wpp.manual.yml).
-    - To create an explorer, we use the custom-made class ExplorerCreator, which has a function `create`. While these object/functions are custom (they combine ds and ds_full tables in a particular way), some of its logic could be generalized and moved to etl.collections. For more details, please refer to the module utils.py.
+    - To create an explorer, we use the custom-made class ExplorerCreator, which has a function `create`. While these object/functions are custom (they combine ds and ds_full tables in a particular way), some of its logic could be generalized and moved to etl.collection. For more details, please refer to the module utils.py.
     - All the created explorers are combined into a single one, which is then exported.
 
 NOTE: This pipeline assumes that there is a TSV template in owid-content, this should probably be change din the future.
 """
 
-from etl.collections.beta import combine_explorers
+from etl.collection.beta import combine_explorers
 from etl.helpers import PathFinder
 
 from .utils import ExplorerCreator
