+++ etl/steps/export/multidim/health/latest/causes_of_death.py
@@ -1,4 +1,4 @@
-from etl.collections import multidim
+from etl.collection import multidim
 from etl.helpers import PathFinder
 
 # Get paths and naming conventions for current step.
