+++ tests/test_collections.py
@@ -1,7 +1,7 @@
 import pytest
 
-from etl.collections.model import CommonView, Definitions
-from etl.collections.utils import merge_common_metadata_by_dimension
+from etl.collection.model import CommonView, Definitions
+from etl.collection.utils import merge_common_metadata_by_dimension
 
 
 def test_merge_common_metadata_1():
